Add files via upload
STAGE 1

DESCRIPTION:
In this project, we will preprocess the data to prepare it for use in a machine learning model that predicts the salaries of NBA players.

Machine learning algorithms are capable of processing big numbers. However, they can't handle raw text and special characters. The latter may include currency signs (for example, $) or measurement units (kg, in). Measurements may also be stated in different units (for example, weight in kg/lb), so you have to choose which unit you intend to use in your data. The records for unwanted units will be considered extraneous and removed. Most ML algorithms don't support missing values, so we have to deal with them before passing the data to ML model.

After dealing with missing values and extraneous characters, it's better to parse the features in the appropriate format for further stages. For example, time features are to be parsed as datetime objects, and features that are continuous variables — as floats.

Once you've completed the steps described above, your data is "clean". In this project, you will practice on the dataset on NBA players. We have automated the data download process in the .py file provided to you. However, if that is inconvenient, feel free to download the nba2k-full.csv dataset on your own.

You can have a look at the dataset profile on Kaggle to learn more about it.

OBJECTIVES:

In this first stage, create a function called clean_data that takes the path to the dataset as a parameter. The function must:

Load a DataFrame from the location specified in the path parameter;

Parse the b_day and draft_year features as datetime objects;

Replace the missing values in team feature with "No Team";

Take the height feature in meters, the height feature contains metric and customary units;

Take the weight feature in kg, the weight feature contains metric and customary units;

Remove the extraneous $ character from the salary feature;

Parse the height, weight, and salary features as floats;

Categorize the country feature as "USA" and "Not-USA";

Replace the cells containing "Undrafted" in the draft_round feature with the string "0";

Return the modified DataFrame.

EXAMPLE:

calling the clean_data function with the path:

path = "../Data/nba2k-full.csv"
df = clean_data(path)
print(df[['b_day', 'team', 'height', 'weight', 'country', 'draft_round', 'draft_year', 'salary']].head())

Output:


          b_day	              team	      height	weight	country	draft_round	draft_year	salary
0	1984-12-30	Los Angeles Lakers	2.06	113.4	USA	   1	        2003-01-01	37436858.0
1	1991-06-29	Los Angeles Clippers    2.01	102.1	USA	   1	        2011-01-01	32742000.0
2	1994-12-06	Milwaukee Bucks	        2.11	109.8	Not-USA	   1	        2013-01-01	25842697.0
3	1988-09-29	Brooklyn Nets	        2.08	104.3	USA	   1	        2007-01-01	37199000.0
4	1989-08-26	Houston Rockets	        1.96	99.8	USA	   1	        2009-01-01	38199000.0